We recently discovered we are leaking auth headers to Loggly.
Currently, the surface area of this is still fairly small, but
obviously not ideal. An ideal infosec turnaround would be <48 hours
to fix this - which is why we are opting to prune auth headers from
the error headers attribute.

In the longer term, we would like to look at re-architecting
nodule-logging, but currently the library is untested and the risk of
contagion from making a risky change in that library is HIGH.

So, in this commit, we make sure to look for both Authorization and
authorization keys in the headers object, and if we find them, we use
lodash's omit() function (chosen because it preserves the original
headers object) to remove them.