Consolidate logs at the operation level
Move logging to the operation level and consolidate with retry logs,
with the overall success (or not) of the operation headlining.

Retry information is included as structured properties of the log line.

This makes it easier to find requests that ended in end-client failure
while still allowing analysis of whether retry configuration is needed
or succesful.